secrets: rename vault to vaultsecrets
