+++ allmodules
@@ -21,4 +21,4 @@ pubsub/natspubsub            yes
 pubsub/rabbitpubsub          yes
 runtimevar/etcdvar           yes
 samples                      no
-secrets/vault                yes
+secrets/vaultsecrets         yes
