+++ internal/testing/alldeps
@@ -77,7 +77,7 @@ gocloud.dev/pubsub/natspubsub
 gocloud.dev/pubsub/rabbitpubsub
 gocloud.dev/runtimevar/etcdvar
 gocloud.dev/samples
-gocloud.dev/secrets/vault
+gocloud.dev/secrets/vaultsecrets
 golang.org/x/crypto
 golang.org/x/net
 golang.org/x/oauth2
