+++ internal/website/content/ref/secrets.md
@@ -23,7 +23,7 @@ Top-level package documentation: <https://godoc.org/gocloud.dev/secrets>
 
 * [Google Cloud KMS](https://godoc.org/gocloud.dev/secrets/gcpkms)
 * [AWS KMS](https://godoc.org/gocloud.dev/secrets/awskms)
-* [Vault by HashiCorp](https://godoc.org/gocloud.dev/secrets/vault) - a
+* [Vault by HashiCorp](https://godoc.org/gocloud.dev/secrets/vaultsecrets) - a
   platform-agnostic secrets engine
 * [In-memory local secrets](https://godoc.org/gocloud.dev/secrets/localsecrets) -
   mainly useful for local testing
