+++ secrets/vaultsecrets/go.mod
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-module gocloud.dev/secrets/vault
+module gocloud.dev/secrets/vaultsecrets
 
 require (
 	github.com/hashicorp/vault/api v1.0.2
