+++ secrets/vaultsecrets/vault.go
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limtations under the License.
 
-// Package vault provides a secrets implementation using the Transit Secrets
+// Package vaultsecrets provides a secrets implementation using the Transit Secrets
 // Engine of Vault by Hashicorp.
 // Use OpenKeeper to construct a *secrets.Keeper.
 //
 // URLs
 //
-// For secrets.OpenKeeper, vault registers for the scheme "vault".
+// For secrets.OpenKeeper, vaultsecrets registers for the scheme "vault".
 // The default URL opener will dial a Vault server using the environment
 // variables "VAULT_SERVER_URL" and "VAULT_SERVER_TOKEN".
 // To customize the URL opener, or for more details on the URL format,
@@ -27,8 +27,8 @@
 //
 // As
 //
-// vault does not support any types for As.
-package vault
+// vaultsecrets does not support any types for As.
+package vaultsecrets
 
 import (
 	"context"
